requirements.txt
# Data processing
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.0  # For Excel file handling

# Machine Learning
scikit-learn>=1.3.0
transformers==4.37.2
torch>=2.0.0
sentence-transformers>=2.2.0
joblib>=1.2.0  # For model serialization
scipy>=1.9.0

# NLP and text processing
nltk>=3.8.0
spacy>=3.5.0

# LLM and Generation
peft==0.10.0
anthropic>=0.5.0

# API clients (optional for real LLM usage)
openai>=1.0.0
anthropic>=0.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Logging
loguru>=0.7.0

# API and deployment
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Utilities
click>=8.1.0
requests>=2.28.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.19.0